+++ CHANGELOG.md
@@ -2,15 +2,16 @@
 
 All significant notes for this project will be documented in this file.
 
-## [Unreleased]
+## [0.0.6] - 2024-12-15
+- Vacation days can be specified in `vacations.yaml`.
+- Each employee is required to take one weekend off per month.
+- Code lines have been moved into functions to enhance readability.
 
 ## [0.0.5] - 2024-12-08
-### Added
 - Refactor.
 - Added `planning.py` script.
 
 ## [0.0.4] - 2023-12-06
-### Added
 - Initial planning functionality.
 - Data export to Excel.
 - Table transposition.